AWSLambda/requirements.txt
boto3
joblib==1.3.2
mecab-python3==1.0.8
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
unidic-lite==1.0.8